refactor(hooks): rename verify_before_complete_hook off the gate prefix
`verify_` is one of the gate-shaped prefixes in `_lib/gate_population.ts`, so
this observability hook sat inside the gate population, where the scan-scope
ratchet counted it as a gate that cannot assert a scan scope. It scans no
corpus at all — its only inputs are its stdin envelope and the state file it
writes — and its contract pins `fail_closed: false` with exit code always 0.

The obvious alternative was excluding `_hook.ts` in the population filter, next
to the existing `.d.ts` / `.test.ts` exclusions. The AI council rejected that as
population-shrinking, and the distinction is real: `.d.ts` and `.test.ts` are
excluded because the language and the test runner make them structurally
non-executable, whereas "a hook never blocks" is an operational property you can
only learn by reading `hook_manifest.yaml` — and manifest-driven classification
was already rejected once, because a new gate lands before anyone registers it.

Renaming keeps classification structural and needs no exclusion rule. The
`_hook.ts` suffix convention is kept (21 of 21 hooks follow it); only the
gate-shaped prefix goes. `verify_physical_move`, the one real `verify_*` gate,
is unaffected and stays in the population.

Also drops the dead golden-parity layer in the moved test (TS_SCRIPT, TSX_BIN,
RunResult, normalize, REPO_ROOT and its now-unused import) — unreferenced before
this change, and own-orphans of it once the file was touched.
